Add remote datacenters to configuration, initialize seeds resource with that
