+++ frameworks/cassandra/src/main/java/com/mesosphere/sdk/cassandra/scheduler/CassandraService.java
@@ -8,8 +8,9 @@ import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 /**
  * Cassandra Service.
@@ -24,10 +25,16 @@ public class CassandraService extends DefaultService {
     @Override
     protected void startApiServer(DefaultScheduler scheduler, int apiPort, Collection<Object> additionalResources) {
         final Collection<Object> apiResources = new ArrayList<>();
+        int configuredSeedCount = Integer.parseInt(System.getenv("SEED_COUNT"));
+        int remoteDatacenterCount = Integer.parseInt(System.getenv("REMOTE_DATA_CENTER_COUNT"));
+        Collection<String> datacenterNames = IntStream.range(0, remoteDatacenterCount).boxed()
+                .map(i -> String.format("dc%d", i))
+                .collect(Collectors.toList());
 
         apiResources.add(
-                new SeedsResource(scheduler.getStateStore(), Arrays.asList("dc2"), 2));
+                new SeedsResource(scheduler.getStateStore(), datacenterNames, configuredSeedCount));
 
+        LOGGER.info("Registering remote datacenters: {}", datacenterNames);
         LOGGER.info("Starting API server with additional resources for Cassandra: {}", apiResources);
         apiResources.addAll(additionalResources);
