+++ frameworks/cassandra/universe/config.json
@@ -45,6 +45,11 @@
           "type": "string",
           "default": "rac1"
         },
+        "seed_count": {
+          "description": "The number of seed nodes to maintain inside a cluster",
+          "type": "integer",
+          "default": 2
+        },
         "remote_data_center_urls": {
           "description": "A list of URLs for schedulers in other data centers for multi-DC clusters",
           "type": "array",
